Fold WooCommerce install/uninstall helper into newfold.mjs
newfold.mjs already owns third-party plugin integration helpers
(supportsWoo, PLUGIN_REQUIREMENTS, etc.), so installWooCommerce/
isWooCommerceActive/uninstallWooCommerce belong there instead of a
separate woocommerce.mjs. Reuses the wordpress/utils helpers newfold.mjs
already imports rather than adding new ones.
